requirements.txt
google-auth>=2.20.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0
google-api-python-client>=2.80.0
pandas>=2.0.0
requests>=2.31.0
tqdm>=4.66.0
